Do not dispatch input callback synchronously because DeckLink API locks the input streams while the callback is in progress.
This can cause a deadlock when calling PauseStreams() (to pause the stream) or StopStreams() at the same time. This happens because we are dispatching all DeckLink API calls to the same serial queue as the input callback. So if we already are on the queue to call PauseStreams() or StopStreams() while a new callback arrives, the callback gets queued up and blocks. At the same time PauseStreams() or StopStreams() waits on the queue for a DeckLink-internal lock that is held by the input callback.